Moves many files around and now it passes
